==> scripts/07-install_packages.R <==
#### Preamble ####
# Purpose: Install necessary packages.
# Author: Yuanyi (Leo) Liu, Dezhen Chen, Ziyuan Shen
# Date: 30 October 2024
# Contact: leoy.liu@mail.utoronto.ca, dezhen.chen@mail.utoronto.ca, ziyuan.shen@mail.utoronto.ca
# License: MIT
# Pre-requisites: None
# Any other information needed? Make sure you are in the `Forecasting-the-2024-US-Presidential-Election` rproj


#### Install packages ####
install.packages("tidyverse")
install.packages("arrow")
install.packages("here")
install.packages("kableExtra")
install.packages("janitor")
install.packages("rstanarm")
install.packages("usmap")
install.packages("knitr")
install.packages("patchwork")
install.packages("car")
install.packages("testthat")
install.packages("rsample")
install.packages("ggplot2")
install.packages("modelsummary")
install.packages("usmap")
